Fix source map errors
